pyproject.toml
[project]
name = "algo-trader"
version = "0.1.0"
description = "Application for managing a complete trading system."
authors = [{ name = "Raimondo Marino", email = "raimondo.marino69@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ibapi",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "setuptools>=80.10.1",
    # PyTorch CUDA 12.6 wheels (driver provides CUDA runtime compatibility).
    "torch==2.10.0",
    "torchvision==0.25.0",
    "torchaudio==2.10.0",
    "pyro-ppl>=1.9.1",
    "ta-lib==0.6.8",
    "ray[default]>=2.53.0",
]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["algo_trader"]

[project.scripts]
algotrader = "algo_trader.main:main"

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.3.260113",
    "pylint>=4.0.4",
    "pyright>=1.1.408",
    "pytest>=7.4",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.uv.sources]
ibapi = { path = "../../IBKR/twsapi_macunix.1042.01/IBJts/source/pythonclient" }
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.pylint.typecheck]
generated-members = ["talib.*"]
